requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.3
django-heroku==0.3.1
gunicorn==20.1.0
idna==3.4
imgurpython==1.1.7
mysqlclient==2.1.1
Pillow==9.3.0
psycopg2==2.9.5
pyimgur==0.6.0
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
whitenoise==6.2.0
